+++ packages/rtp/src/extra/processor/rtpTime.ts
@@ -1,4 +1,4 @@
-import { int, RtpPacket } from "..";
+import { int, RtpPacket } from "../..";
 import { Processor } from "./interface";
 import { Max32Uint } from "./webm";
